Add infrastructure needed to interpret, convert and search time.
What this change does is record the temporal units meta information
from the netcdfcf file into the database.  Now we can convert the
unitless timestamps given by netcdf into an absolute "days since
0-0-0" format (for comparison) and into Y-M-D format to present
to the user.

The database now records: the timesteps (as before), the units
(so you can convert at will) and the min and max times in absolute
and YMD formats.
NOTES:
1) The UI is not yet changed and looks and works just like before.
2) The converter python module functions OK now now but it should
be converted into an interface into cfunits for better fidelity.
3) Requires a corresponding VTK change to get the units back.