Improving legend API
Adding ability to set legend position and size for a given layer. Basically
the idea is that the view can have multiple layers each with its own legend.Now
if multiple layers are visible then we want legends not to overlap. Now the
view can find positions and sizes for each layer and run some space resolution
algorithm to make sure that they don't overlap