Add session authentication using ESGF OpenIds
Add custom SessionAuth tool to use MyProxyLogon to authenticate
user at login. The cert is then saved for later use in downloaded
ESGF datasets. The first time a user login an entry in the 'users'
collection is create in mongo recording the OpenId, this can be
used for user specfic configuration in the future.